+++ lib/src/widgets/pin.widget.dart
@@ -45,7 +45,6 @@ class _PinWidgetState extends State<PinWidget> {
               textAlign: TextAlign.center,
               style: TextStyle(
                 fontWeight: FontWeight.w500,
-                color: Colors.black87,
                 fontSize: 15.0,
               ),
             ),
@@ -55,7 +54,6 @@ class _PinWidgetState extends State<PinWidget> {
               textAlign: TextAlign.center,
               style: TextStyle(
                 fontWeight: FontWeight.bold,
-                color: Colors.black87,
                 fontSize: 25.0,
                 letterSpacing: 15.0,
               ),